add margin account supervisor force withdrawal

+++ user-service/src/main/java/rs/edu/raf/banka1/services/implementations/MarginAccountServiceImpl.java
@@ -11,16 +11,12 @@ import rs.edu.raf.banka1.margincalljob.MarginCallMidnightJob;
 import rs.edu.raf.banka1.margincalljob.MarginCallMidnightTrigger;
 import rs.edu.raf.banka1.model.ListingType;
 import rs.edu.raf.banka1.model.MarginAccount;
-import rs.edu.raf.banka1.model.*;
 import org.tinylog.Logger;
 import rs.edu.raf.banka1.dtos.MarginAccountDto;
 import rs.edu.raf.banka1.model.BankAccount;
 import rs.edu.raf.banka1.model.Customer;
 import rs.edu.raf.banka1.repositories.MarginAccountRepository;
-import rs.edu.raf.banka1.services.BankAccountService;
-import rs.edu.raf.banka1.services.EmailService;
-import rs.edu.raf.banka1.services.MarginAccountService;
-import rs.edu.raf.banka1.services.MarginTransactionService;
+import rs.edu.raf.banka1.services.*;
 import rs.edu.raf.banka1.utils.Constants;
 
 import java.util.ArrayList;
@@ -118,9 +114,16 @@ public class MarginAccountServiceImpl implements MarginAccountService {
             return false;
         }
 
+        if(amount > marginAccount.getMaintenanceMargin() - marginAccount.getBalance()) {
+            Logger.error("Wrong amount for deposit margin call.");
+            return false;
+        }
+
         bankAccountService.removeBalance(bankAccount, amount);
         depositToMarginAccount(marginAccount, amount);
         marginTransactionService.createTransactionMarginCall(marginAccount, amount);
+        marginAccount.setMarginCallLevel(0);
+        marginAccountRepository.save(marginAccount);
         return true;
     }
 
@@ -185,5 +188,16 @@ public class MarginAccountServiceImpl implements MarginAccountService {
         }
         marginAccount.setMarginCallLevel(2);
         this.marginAccountRepository.save(marginAccount);
+    }
+
+    @Override
+    public Boolean supervisorForceWithdrawal(Long marginAccountId) {
+        MarginAccount marginAccount = marginAccountRepository.findById(marginAccountId).orElseThrow(() -> new MarginAccountNotFoundException(marginAccountId, null, null));
+        if(marginAccount.getMarginCallLevel() != 2) {
+            Logger.error("Margin Call Level is not 2.");
+            return false;
+        }
+        Double amount = marginAccount.getMaintenanceMargin() - marginAccount.getBalance();
+        return depositMarginCall(marginAccountId, amount);
     }
 }
